Version bump and add apache, docker, haproxy

+++ bin/install-docker.sh
@@ -2,17 +2,19 @@
 set -e
 #set -x
 
-echo "--> Adding docker keyserver"
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add - &>/dev/null
+if [[ ! -f $(which docker) ]]; then
+  echo "--> Adding docker keyserver"
+  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add - &>/dev/null
 
-echo "--> Adding docker repo"
-sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
+  echo "--> Adding docker repo"
+  sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
 
-echo "--> Updating cache"
-sudo apt-get update --fix-missing
+  echo "--> Updating cache"
+  sudo apt-get update --fix-missing
 
-echo "--> Installing docker"
-sudo apt-get install docker-ce --yes
+  echo "--> Installing docker"
+  sudo apt-get install docker-ce --yes
 
-echo "--> Allowing docker without sudo"
-sudo usermod -aG docker "$(whoami)"
+  echo "--> Allowing docker without sudo"
+  sudo usermod -aG docker "$(whoami)"
+fi
